pyproject.toml
[project]
name = "eventizer-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "chromadb>=1.5.1",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-genai>=1.64.0",
    "mysqlclient>=2.2.8",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "redis>=7.3.0",
    "requests>=2.32.5",
]
[tool.black]
line-length = 88
target-version = ['py313']
exclude = '''
/(
    \.git
  | \.venv
  | \.env
  | \.env-docker
  | build
  | dist
)/
'''
